Fix signature check scoping issue.
Signature checks were failing, in the binary-document case of
openpgp_packet_signature.verify() the variable i was being referenced
when there was no i in scope; instead it was getting 8 from a global
variable i, and trimming 8 bytes off the beginning of the data.

The global i was being set by the CAST5 cipher -- this also fixes
that to be a local variable.

Also add unit test to confirm.