Properly encode length of key material in v5
When private keys are serialized unencrypted, a 2-byte checksum is
appended after the actual key material. According to spec, these 2
bytes are not included in the length of the key material. The openpgpjs
implementation, however, erroneously includes them.

Since version 5 packets contain this length, which is used when they
are parsed, correct implementation of the specification will not accept
secret keys from openpgpjs, when they are serialized using version 5
packets.

To fix this, calculate the length of the key material before appending
the checksum

Fixes #1155