Update README.md
Changed `openpgp.readKey()` to `openpgp.readPrivateKey()` where applicable. I noticed this documentation error when I was writing my own app in typescript and vscode kept yelling at me that it wanted a PrivateKey object instead of a PublicKey object, and I was confused because I wrote everything correct according to this README. I looked into the type definitions file and found that `readKey()` only returns a PublicKey object, and `readPrivateKey()` only returns a PrivateKey object, which is what was needed for `decryptKey()`.