Fix CFB decryption performance in JS fallback for ciphers other than AES
By calling the underlying cipher decryption function for one block at a time.
Before, the entire ciphertext was passed for decryption, but only a subset of it (equal to the block size)
was actually processed in each iteration.
This resulted in a lot of extra work for e.g. Cast5 decryption, whose decryption function supported decrypting more than a block at once.

This issue affected non-AES ciphers (legacy), such as Cast5, in Node 18+ or in browser.