Handle backpressure in fflate fallback
Don't start (de)compressing more data if the (de)compressed data stream
reader hasn't asked for more data yet.

Also, split the input chunks so that we don't overwhelm the output
stream with a lot of data if the input chunks are very large, as
fflate seems to decompress the input chunks completely before
calling `ondata`, which doesn't give us a chance to cancel the
decompression if too much data is provided.